# chunks: improve readability of compression benchmarks